Add files via upload
# Petrol price hike analysis
Rising fuel prices shocks have a significant impact on the way of life of most citizens. Due to the rising of oil prices, the costs of travel for private vehicle users are therefore increasing. Petrol is the most used fuel for all the means of transport. At present in every family there is at least one personal vehicle. Hike in the petrol prices now-a-days has been an important issue for every middle class and lower middle class families. In this project we will analyze the petrol prices from the past years using bar graphs, pie charts and other pictorial representations using QlikView tool. From this project we will be able to conclude the amount of hike in the prices at a particular year or several years.

# Supermarket sales analysis
The growth of supermarket in popular cities are increasing and market competitions are also high. Sales forecasting is an important when it comes to companies who are engaged in retailing, logistics, manufacturing, marketing and wholesaling. It allows companies to allocate resources efficiently, to estimate revenue of the sales and to plan strategies which are better for company’s future. Predicting the demand accurately is critical for the organizations in order to be able to formulate the plans. In this project we take dataset of a supermarket sales at three different branches and analyze the prices, the product which is most preffered by the customers by using QlikView tool and analyze the sales by bar graphs and pie charts. Hence, the project “SUPERMARKET SALES ANALYSIS” aims to predict the sales of the mart by which rate of customers increases and profit increases.